refactoring: move TileManager to app
